+++ src/app/content/profile/pages/profile-list-page/profile-list-page.component.ts
@@ -3,7 +3,6 @@ import { faUsers } from '@fortawesome/free-solid-svg-icons';
 import { Country, ProfileItem } from '@/models';
 import { ProfileService, CountryService } from '@/services';
 import { Observable } from 'rxjs';
-import { first } from 'rxjs/operators';
 
 @Component({
   selector: 'app-profile-list-page',
@@ -14,7 +13,7 @@ export class ProfileListPageComponent {
   public profiles$: Observable<ProfileItem[]>;
   public countries: Country[];
 
-  // UI
+  // Font Awesome
   faUsers = faUsers;
 
   /**
@@ -25,32 +24,7 @@ export class ProfileListPageComponent {
    * @memberof ProfileListPageComponent
    */
   constructor(private profileService: ProfileService, private countryService: CountryService) {
-    this.getProfiles();
-    this.getCountries();
-  }
-
-  /**
-   * Get the profiles's list
-   *
-   * @private
-   * @memberof ProfileListPageComponent
-   */
-  private getProfiles(): void {
     this.profiles$ = this.profileService.getProfiles();
-  }
-
-  /**
-   * Get the list of countries
-   *
-   * @private
-   * @memberof ProfileEditPageComponent
-   */
-  private getCountries(): void {
-    this.countryService
-      .getCountries()
-      .pipe(first())
-      .subscribe((countries: Country[]) => {
-        this.countries = countries;
-      });
+    this.countries = this.countryService.getCountries();
   }
 }